java: fix positional completion
- complete filenames without `.class` suffix (assume these are named
  correctly)
- only complete these if neither classpath flag nor env is set
- support classpath env for classpath classes completion
